🐛 fix: 共通の関数をまとめ、それらをimport
